+++ libcaf_core/caf/scheduled_actor.hpp
@@ -853,11 +853,6 @@ private:
   /// Stores children that were marked for release while running an action.
   std::vector<flow::coordinated_ptr> released_;
 
-  /// Counter for scheduled_actor::delay to make sure
-  /// scheduled_actor::run_actions does not end up in a busy loop that might
-  /// starve other activities.
-  size_t delayed_actions_this_run_ = 0;
-
   /// Stores ongoing activities such as flows that block the actor from
   /// terminating.
   std::vector<disposable> watched_disposables_;